reference_apps/onboarder_github_fastapi/requirements.txt
fastapi[all]
requests
pydantic
pydantic-settings
uvicorn
PyGithub
redis
